+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Fix up types
 
+### Removed
+- Remove `LSFiles.populate_regex`
+
 [0.4.0](https://github.com/jshwi/lsfiles/releases/tag/v0.4.0) - 2023-02-18
 ------------------------------------------------------------------------
 ### Added